ISSUE #2914 - moved destination and revision tag inputs back to individual dirs
