+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadFileForm/uploadList/uploadListItem/components/uploadListItemDestination/uploadListItemDestination.styles.ts
@@ -16,20 +16,14 @@
  */
 
 import styled from 'styled-components';
-import { Typography } from '@controls/typography';
+import { TextField } from '@material-ui/core';
 
-export const Container = styled.div`
-	background-color: ${({ theme }) => theme.palette.tertiary.lightest};
-	color: ${({ theme }) => theme.palette.secondary.main};
-	border-radius: 5px;
-	padding: 9px 12px;
+export const Input = styled(TextField)`
+	margin: 0 7px;
 	width: 271px;
-	margin: 0 10px;
-
-	box-sizing: border-box;
-`;
-
-export const ContainerInput = styled(Typography).attrs({
-	variant: 'h5',
-})`
+	${({ $selectedrow, theme }) => $selectedrow && `
+		div { background-color: ${theme.palette.secondary.light} }
+		&&& input { color: ${theme.palette.primary.contrast} }
+		&& fieldset { border-color: transparent; }
+	`}
 `;
